requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.2.5
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
decorator==4.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.9.4
entrypoints==0.3
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.12.1
Jinja2==2.10.1
json5==0.8.5
jsonpickle==1.2
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.4.0
jupyterlab==1.0.4
jupyterlab-server==1.0.0
lazy-object-proxy==1.4.1
lxml==4.3.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.15.0
pandas==0.23.4
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==1.0.15
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
PyAutoGUI==0.9.48
pycparser==2.19
PyDispatcher==2.0.5
PyGetWindow==0.0.7
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==2.3.1
PyMsgBox==1.0.7
pyobjc==6.0.1
pyobjc-core==6.0.1
pyobjc-framework-Accounts==6.0.1
pyobjc-framework-AddressBook==6.0.1
pyobjc-framework-AdSupport==6.0.1
pyobjc-framework-AppleScriptKit==6.0.1
pyobjc-framework-AppleScriptObjC==6.0.1
pyobjc-framework-ApplicationServices==6.0.1
pyobjc-framework-Automator==6.0.1
pyobjc-framework-AVFoundation==6.0.1
pyobjc-framework-AVKit==6.0.1
pyobjc-framework-BusinessChat==6.0.1
pyobjc-framework-CalendarStore==6.0.1
pyobjc-framework-CFNetwork==6.0.1
pyobjc-framework-CloudKit==6.0.1
pyobjc-framework-Cocoa==6.0.1
pyobjc-framework-Collaboration==6.0.1
pyobjc-framework-ColorSync==6.0.1
pyobjc-framework-Contacts==6.0.1
pyobjc-framework-ContactsUI==6.0.1
pyobjc-framework-CoreAudio==6.0.1
pyobjc-framework-CoreAudioKit==6.0.1
pyobjc-framework-CoreBluetooth==6.0.1
pyobjc-framework-CoreData==6.0.1
pyobjc-framework-CoreLocation==6.0.1
pyobjc-framework-CoreMedia==6.0.1
pyobjc-framework-CoreMediaIO==6.0.1
pyobjc-framework-CoreML==6.0.1
pyobjc-framework-CoreServices==6.0.1
pyobjc-framework-CoreSpotlight==6.0.1
pyobjc-framework-CoreText==6.0.1
pyobjc-framework-CoreWLAN==6.0.1
pyobjc-framework-CryptoTokenKit==6.0.1
pyobjc-framework-DictionaryServices==6.0.1
pyobjc-framework-DiscRecording==6.0.1
pyobjc-framework-DiscRecordingUI==6.0.1
pyobjc-framework-DiskArbitration==6.0.1
pyobjc-framework-DVDPlayback==6.0.1
pyobjc-framework-EventKit==6.0.1
pyobjc-framework-ExceptionHandling==6.0.1
pyobjc-framework-ExternalAccessory==6.0.1
pyobjc-framework-FinderSync==6.0.1
pyobjc-framework-FSEvents==6.0.1
pyobjc-framework-GameCenter==6.0.1
pyobjc-framework-GameController==6.0.1
pyobjc-framework-GameKit==6.0.1
pyobjc-framework-GameplayKit==6.0.1
pyobjc-framework-ImageCaptureCore==6.0.1
pyobjc-framework-IMServicePlugIn==6.0.1
pyobjc-framework-InputMethodKit==6.0.1
pyobjc-framework-InstallerPlugins==6.0.1
pyobjc-framework-InstantMessage==6.0.1
pyobjc-framework-Intents==6.0.1
pyobjc-framework-IOSurface==6.0.1
pyobjc-framework-iTunesLibrary==6.0.1
pyobjc-framework-LatentSemanticMapping==6.0.1
pyobjc-framework-LaunchServices==6.0.1
pyobjc-framework-libdispatch==6.0.1
pyobjc-framework-LocalAuthentication==6.0.1
pyobjc-framework-MapKit==6.0.1
pyobjc-framework-MediaAccessibility==6.0.1
pyobjc-framework-MediaLibrary==6.0.1
pyobjc-framework-MediaPlayer==6.0.1
pyobjc-framework-MediaToolbox==6.0.1
pyobjc-framework-MetalKit==6.0.1
pyobjc-framework-ModelIO==6.0.1
pyobjc-framework-MultipeerConnectivity==6.0.1
pyobjc-framework-NaturalLanguage==6.0.1
pyobjc-framework-NetFS==6.0.1
pyobjc-framework-Network==6.0.1
pyobjc-framework-NetworkExtension==6.0.1
pyobjc-framework-NotificationCenter==6.0.1
pyobjc-framework-OpenDirectory==6.0.1
pyobjc-framework-OSAKit==6.0.1
pyobjc-framework-Photos==6.0.1
pyobjc-framework-PhotosUI==6.0.1
pyobjc-framework-PreferencePanes==6.0.1
pyobjc-framework-PubSub==6.0.1
pyobjc-framework-QTKit==6.0.1
pyobjc-framework-Quartz==6.0.1
pyobjc-framework-SafariServices==6.0.1
pyobjc-framework-SceneKit==6.0.1
pyobjc-framework-ScreenSaver==6.0.1
pyobjc-framework-ScriptingBridge==6.0.1
pyobjc-framework-SearchKit==6.0.1
pyobjc-framework-Security==6.0.1
pyobjc-framework-SecurityFoundation==6.0.1
pyobjc-framework-SecurityInterface==6.0.1
pyobjc-framework-ServiceManagement==6.0.1
pyobjc-framework-Social==6.0.1
pyobjc-framework-SpriteKit==6.0.1
pyobjc-framework-StoreKit==6.0.1
pyobjc-framework-SyncServices==6.0.1
pyobjc-framework-SystemConfiguration==6.0.1
pyobjc-framework-UserNotifications==6.0.1
pyobjc-framework-VideoSubscriberAccount==6.0.1
pyobjc-framework-VideoToolbox==6.0.1
pyobjc-framework-Vision==6.0.1
pyobjc-framework-WebKit==6.0.1
pyOpenSSL==18.0.0
pyparsing==2.4.0
pyperclip==1.7.0
PyRect==0.1.4
pyrsistent==0.15.4
PyScreeze==0.1.24
python-dateutil==2.7.3
python-dotenv==0.10.3
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.1
pyzmq==17.1.0
qiniu==7.2.6
queuelib==1.5.0
requests==2.22.0
scikit-learn==0.19.2
scipy==1.1.0
Scrapy==1.5.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
svgwrite==1.3.0
terminado==0.8.2
testpath==0.4.2
tornado==5.1
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.6.1
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.3
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
zope.interface==4.6.0
